fix clang build breakage
Summary: ^

Test Plan: compiled.  What a concept.

Reviewed By: woo@fb.com

Subscribers: doug, bmatheny, mathieubaudet

FB internal diff: D1756489

Signature: t1:1756489:1419369051:87194d17f843b3f71f5aa1dda85a2e141fe07b1a

+++ proxygen/lib/http/session/HTTPSession.h
@@ -227,14 +227,14 @@ class HTTPSession:
   void setMaxConcurrentPushTransactions(uint32_t num);
 
   /**
-   * Get the number of egress bytes this session will buffer before
+   * Get/Set the number of egress bytes this session will buffer before
    * pausing all transactions' egress.
    */
-  static uint64_t getPendingWriteMax() {
+  static uint32_t getPendingWriteMax() {
     return kPendingWriteMax;
   }
 
-  static void setPendingWriteMax(uint64_t max) {
+  static void setPendingWriteMax(uint32_t max) {
     kPendingWriteMax = max;
   }
